requirements.txt
# RunPod SDK
runpod>=1.6.0

# Coqui TTS XTTS_v2 (qualité professionnelle)
TTS==0.22.0
transformers==4.33.0

# PyTorch pour GPU (version compatible avec Coqui TTS)
torch==2.1.2
torchaudio==2.1.2

# Traitement basique
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Wav2Lip pour génération vidéo
opencv-python>=4.8.0
librosa>=0.10.0
face-alignment>=1.3.5
scipy>=1.11.0
numba==0.57.0
